fixes flakiness of the dirty ready test
ensures that the ES client does not try to set the number of processors
when the client initializes. it could be already set at inside the same
JVM which raises an exception.
although this setting is just an upper limit to prevent too much threads
on machines with lot of cores (>32), this is not relevant for our tests
here.