Avoid "availableProcessors is already set" exception.
Ensures that the ES client does not try to set the number of processors
when the client initializes. it could be already set at inside the same
JVM which raises an exception.
Although this setting is just an upper limit to prevent too many threads
on machines with lot of cores (>32), this is not relevant for our tests
here.

https://github.com/netty/netty/issues/6956