Disable the column store for the sets table.
We use a generate method to insert values into the sets table (the elements
column) which can be larger than the 32766 bytes limit the Lucene enforces
per term. Since we only query by id in our test we disable the index on
the elements column so we can insert tha arbitrary large values.